fix: selected state check, deduplicate getFileColor, fix Playwright tests
- Fix isSelected check: directory nodes should not be selected when selectedFile is null
- Extract getFileColor to helpers.ts, use in both SkillFileBrowser and SkillFilePreview
- Rewrite Playwright tests to match actual UI (button opens Modal, not tab)
